Docs, Swagger and Postman Collection

+++ src/main/java/com/brecom/ecommerce/inventories/orders/application/services/DeleteOneOrderByIdService.java
@@ -2,6 +2,7 @@ package com.brecom.ecommerce.inventories.orders.application.services;
 
 import com.brecom.ecommerce.inventories.orders.application.ports.in.DeleteOneOrderByIdUseCase;
 import com.brecom.ecommerce.inventories.orders.application.ports.out.OrderPersistencePort;
+import com.brecom.ecommerce.inventories.orders.domain.exceptions.OrderNotFoundException;
 
 import org.springframework.stereotype.Service;
 
@@ -15,6 +16,12 @@ public class DeleteOneOrderByIdService implements DeleteOneOrderByIdUseCase {
     }
     @Override
     public Mono<Void> execute(Integer id) {
-        return this.orderPersistencePort.deleteOneOrderById(id);
+        return this.orderPersistencePort.existsOneOrderById(id)
+                .flatMap(exists -> {
+                    if (!exists) {
+                        return Mono.error(new OrderNotFoundException("Order with id '" + id + "' not found"));
+                    }
+                    return this.orderPersistencePort.deleteOneOrderById(id);
+                });
     }
 }
